# app/artifact_agent/app/src/document/yml/openapi.yml
openapi: 3.0.3
info:
  title: Create Ingestion Settings API
  version: 1.0.0
servers:
- url: http://api.example.com/v1
paths:
  /ingestion-settings:
    post:
      summary: Create Ingestion Setting
      operationId: createIngestionSetting
      tags:
      - IngestionSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionSettingCreateRequest'
      responses:
        '201':
          description: Ingestion Setting created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionSettingResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
    put:
      summary: Update Ingestion Setting
      operationId: updateIngestionSetting
      tags:
      - IngestionSettings
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestionSettingUpdateRequest'
      responses:
        '200':
          description: Ingestion Setting updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionSettingResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Ingestion Setting not found
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
    delete:
      summary: Delete Ingestion Setting
      operationId: deleteIngestionSetting
      tags:
      - IngestionSettings
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Ingestion Setting deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Ingestion Setting not found
        '500':
          description: Internal Server Error
components:
  schemas:
    IngestionSettingCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    IngestionSettingUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    IngestionSettingResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
x-backend: .net core
x-database: mongodb