# app/artifact_agent/app/src/generated_outputs/.net_core/create_ingestion_settings_api/IngestionSettingsApi/Controllers/IngestionSettingsController.cs
using Microsoft.AspNetCore.Mvc;
using System.Threading.Tasks;

namespace IngestionSettingsApi.Controllers
{
    [ApiController]
    [Route("[controller]")]
    public class IngestionSettingsController : ControllerBase
    {
        [HttpPost]
        public async Task<IActionResult> CreateIngestionSetting([FromBody] IngestionSettingCreateRequest request)
        {
            // Implementation here
            return CreatedAtAction(nameof(GetIngestionSetting), new { id = "newId" }, new IngestionSettingResponse());
        }

        [HttpPut("{id}")]
        public async Task<IActionResult> UpdateIngestionSetting(string id, [FromBody] IngestionSettingUpdateRequest request)
        {
            // Implementation here
            return Ok(new IngestionSettingResponse());
        }

        [HttpDelete("{id}")]
        public async Task<IActionResult> DeleteIngestionSetting(string id)
        {
            // Implementation here
            return NoContent();
        }

        // Additional methods for getting ingestion settings would go here
    }
}