# app/artifact_agent/app/src/generated_outputs/.net_core/create_ingestion_settings_api/IngestionSettingsApi/Models/IngestionSettingResponse.cs
public class IngestionSettingResponse
{
    public string Id { get; set; }
    public string Name { get; set; }
    public string Description { get; set; }
    public bool IsActive { get; set; }
    public DateTime CreatedAt { get; set; }
    public DateTime UpdatedAt { get; set; }
}