# app/artifact_agent/app/src/generated_outputs/java/calculator_api/src/main/java/com/example/calculatorapi/model/ErrorResponse.java
package com.example.calculatorapi.model;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class ErrorResponse {
    private String error;
    private String message;
    private int statusCode;
}